fix(web): Support user info in logs on web.
`setUserInfo` only set user info on RUM, not Logs This was an unintentional oversight.

Web still does not support extra user info. This will be tracked as a separate issue and may be part of some changes coming in v3.

refs: #781
